requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-environ>=0.11
psycopg2-binary>=2.9
drf-yasg>=1.21
django-cors-headers>=4.3
django-filter>=24.2
django-crispy-forms>=2.3
django-allauth>=0.63
celery>=5.3
redis>=5.0
pytest>=8.0
pytest-django>=4.8
django-debug-toolbar>=4.4
gunicorn>=22.0
